--- api/metric/batch.go
@@ -1,1 +1,0 @@
-package metric
